added tracers
