+++ api/router.go
@@ -4,8 +4,10 @@ import (
 	"context"
 	"delta/config"
 	"delta/core"
+	"delta/utils"
 	"encoding/json"
 	"fmt"
+	"github.com/application-research/delta-db/messaging"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
 	"net/http"
@@ -89,6 +91,36 @@ func InitializeEchoRouterConfig(ln *core.DeltaNode, config config.DeltaConfig) {
 			span.SetAttributes(attribute.String("host", c.Request().Host))
 			span.SetAttributes(attribute.String("referer", c.Request().Referer()))
 			span.SetAttributes(attribute.String("request_uri", c.Request().RequestURI))
+
+			fmt.Println("Request: " + c.Request().Method + " " + c.Path() + " " + c.Request().UserAgent() + " " + c.RealIP() + " " + c.Request().Host + " " + c.Request().Referer() + " " + c.Request().RequestURI)
+			s := struct {
+				RemoteIP string `json:"remote_ip"`
+				Host     string `json:"host"`
+				Referer  string `json:"referer"`
+				Request  string `json:"request"`
+				Path     string `json:"path"`
+			}{
+				RemoteIP: c.RealIP(),
+				Host:     c.Request().Host,
+				Referer:  c.Request().Referer(),
+				Request:  c.Request().RequestURI,
+				Path:     c.Path(),
+			}
+
+			b, err := json.Marshal(s)
+			if err != nil {
+				log.Error(err)
+			}
+
+			utils.GlobalDeltaDataReporter.TraceLog(
+				messaging.LogEvent{
+					LogEventType:   "Route Request",
+					LogEventObject: b,
+					LogEvent:       string(b),
+					CreatedAt:      time.Now(),
+					UpdatedAt:      time.Now(),
+				})
+
 			return next(c)
 		}
